test(sync): isolate estimate preauthorization coverage
